requirements.txt
sounddevice
numpy
openai
git+https://github.com/openai/whisper.git
tqdm
python-dotenv
